+++ tools/perf_vector_svml_expression.c
@@ -155,7 +155,7 @@ int main(int argc, char** argv)
         config.filter_flags = IARRAY_COMP_SHUFFLE;
         if (mantissa_bits > 0) {
             config.filter_flags |= (int) IARRAY_COMP_TRUNC_PREC;
-            config.fp_mantissa_bits = mantissa_bits;
+            config.fp_mantissa_bits = (uint8_t) mantissa_bits;
         }
     }
     config.use_dict = INA_SUCCEED(ina_opt_isset("d")) ? 1 : 0;
